# src/main/java/com/spring/cart/demo/api/controller/CartsController.java
package com.spring.cart.demo.api.controller;

import static org.springframework.http.ResponseEntity.ok;

import com.spring.cart.demo.api.model.Cart;
import com.spring.cart.demo.api.model.Item;
import java.util.Collections;
import java.util.List;

import jakarta.validation.Valid;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class CartsController  {

    private static final Logger log = LoggerFactory.getLogger(CartsController.class);

    public ResponseEntity<List<Item>> addCartItemsByCustomerId(String customerId, @Valid Item item) {
        log.info("Request for customer ID: {}\nItem: {}", customerId, item);
        return ok(Collections.EMPTY_LIST);
    }


    public ResponseEntity<List<Item>> addOrReplaceItemsByCustomerId(String customerId,
                                                                    @Valid Item item) {
        return null;
    }


    public ResponseEntity<Void> deleteCart(String customerId) {
        return null;
    }


    public ResponseEntity<Void> deleteItemFromCart(String customerId, String itemId) {
        return null;
    }


    public ResponseEntity<List<Cart>> getCartByCustomerId(String customerId) {
        throw new RuntimeException("Manual Exception thrown");
    }


    public ResponseEntity<List<Item>> getCartItemsByCustomerId(String customerId) {
        return null;
    }


    public ResponseEntity<List<Item>> getCartItemsByItemId(String customerId, String itemId) {
        return null;
    }
}